requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dependency-injector==4.41.0
Deprecated==1.2.14
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.1
multidict==6.0.4
numpy==1.25.0
opencv-python==4.8.0.74
opentelemetry-api==1.18.0
packaging==23.1
Pillow==10.0.0
pydantic==2.0.1
pydantic-settings==2.0.1
pydantic_core==2.0.2
pyTelegramBotAPI==4.12.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tesseract==0.1.3
typing_extensions==4.7.1
urllib3==2.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.0
